frontend/requirements.txt
# Core UI
streamlit>=1.32.0

# Already in your project (adjust versions to match yours)
python-dotenv>=1.0.0

# Add your existing deps below — these are placeholders
# openai>=1.0.0
# langchain>=0.2.0
# langchain-community>=0.2.0
# yt-dlp>=2024.1.0
# faster-whisper>=1.0.0
# chromadb>=0.5.0
